Improves the tests for encoders
