+++ encodings/encoding-binary/src/test/java/FixedEncoderDecoderTest.java
@@ -22,24 +22,20 @@
 import esa.mo.mal.encoder.binary.fixed.FixedBinaryStreamFactory;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
+import org.ccsds.moims.mo.mal.MALException;
+import org.ccsds.moims.mo.mal.OperationField;
 import org.ccsds.moims.mo.mal.encoding.MALElementInputStream;
 import org.ccsds.moims.mo.mal.encoding.MALElementOutputStream;
-import org.ccsds.moims.mo.mal.structures.Identifier;
-import org.ccsds.moims.mo.mal.structures.NullableAttribute;
-import org.ccsds.moims.mo.mal.structures.NullableAttributeList;
-import org.ccsds.moims.mo.mal.structures.StringList;
+import org.ccsds.moims.mo.mal.structures.Element;
 import org.junit.After;
 import org.junit.AfterClass;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Test;
 
 /**
  * Tests for Encoding and Decoding with FixedBinary
  */
-public class FixedEncoderDecoderTest {
+public class FixedEncoderDecoderTest extends EncoderDecoderTest {
 
     public FixedEncoderDecoderTest() {
     }
@@ -60,69 +56,18 @@ public class FixedEncoderDecoderTest {
     public void tearDown() {
     }
 
-    /**
-     * Simple test for encoding decoding a list of NullableAttributeList.
-     *
-     * @throws java.lang.Exception
-     */
-    @Test
-    public void testEncodeDecodeNullableAttributeList() throws Exception {
-        NullableAttributeList list = new NullableAttributeList();
-        list.add(new NullableAttribute(new Identifier("String A")));
-        list.add(new NullableAttribute(new Identifier("String B")));
-        list.add(new NullableAttribute(new Identifier("String C")));
-
+    @Override
+    public Element encodeThenDecode(Element element, OperationField field) throws MALException {
+        // Encode
         FixedBinaryStreamFactory factory = new FixedBinaryStreamFactory();
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         MALElementOutputStream malWriter = factory.createOutputStream(baos);
+        malWriter.writeElement(element, field);
 
-        malWriter.writeElement(list, null);
-
+        // Decode
         final ByteArrayInputStream bais = new ByteArrayInputStream(baos.toByteArray());
         MALElementInputStream malReader = factory.createInputStream(bais);
-
-        NullableAttributeList readList = new NullableAttributeList();
-        readList = (NullableAttributeList) malReader.readElement(readList, null);
-
-        // Assertions:
-        assertNotNull(readList);
-        assertEquals(3, readList.size());
-        assertEquals("NullableAttributeList.get(0)", list.get(0), readList.get(0));
-        assertEquals("NullableAttributeList.get(1)", list.get(1), readList.get(1));
-        assertEquals("NullableAttributeList.get(2)", list.get(2), readList.get(2));
+        return malReader.readElement(element.createElement(), field);
     }
 
-    /**
-     * Simple test for encoding decoding a list of StringList. Note that the
-     * String MAL type is mapped to the String Java type, therefore the type
-     * might need to be wrapped in a Union type for correct encoding/decoding.
-     *
-     * @throws java.lang.Exception
-     */
-    @Test
-    public void testEncodeDecodeStringList() throws Exception {
-        StringList list = new StringList();
-        list.add("String A");
-        list.add("String B");
-        list.add("String C");
-
-        FixedBinaryStreamFactory factory = new FixedBinaryStreamFactory();
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        MALElementOutputStream malWriter = factory.createOutputStream(baos);
-
-        malWriter.writeElement(list, null);
-
-        final ByteArrayInputStream bais = new ByteArrayInputStream(baos.toByteArray());
-        MALElementInputStream malReader = factory.createInputStream(bais);
-
-        StringList readList = new StringList();
-        readList = (StringList) malReader.readElement(readList, null);
-
-        // Assertions:
-        assertNotNull(readList);
-        assertEquals(3, readList.size());
-        assertEquals("StringList.get(0)", list.get(0), readList.get(0));
-        assertEquals("StringList.get(1)", list.get(1), readList.get(1));
-        assertEquals("StringList.get(2)", list.get(2), readList.get(2));
-    }
 }
